🎨 UserInfo 컴포넌트를 commons/UserChip으로 분리
