+++ app/frontend/src/components/commons/Chatting/TalkItem.css.ts
@@ -50,30 +50,4 @@ export const datetime = style([
   },
 ]);
 
-export const defaultProfileImage = style({
-  border: `1px solid ${vars.color.morakGreen}`,
-  borderRadius: '50%',
-});
-
 export { isMine };
-
-export const profileImage = style({
-  display: 'flex',
-  flexShrink: 0,
-  width: '2.4rem',
-  height: '2.4rem',
-  borderRadius: '50%',
-  overflow: 'hidden',
-});
-
-export const userInfo = style([
-  sansRegular12,
-  {
-    display: 'flex',
-    alignItems: 'center',
-    flexShrink: 0,
-    gap: '0.4rem',
-    overflow: 'hidden',
-    whiteSpace: 'nowrap',
-  },
-]);
